+++ features/step_definitions/partner_steps/fanatics_steps.rb
@@ -30,3 +30,28 @@ Given(/^FANATICS\- I verify the Troop ID discount has been applied$/) do
   expect(find("#checkoutCouponTotalValue").text).to eql("-$12.00")
   expect(page).to have_selector("#desktopIdMeVerified",:text =>"Status Verified",:visible => false)
 end
+
+Given(/^FANATICS_M\- I add an item to the cart$/) do
+  visit FigNewton.partners.fanatics.mobile.product_page
+  find("#pdpSizeSelector div",:match => :first).click
+  find("#mobileAddToCart").click
+  find("#militaryAndFirstResponderHeader").click
+end
+
+Given(/^FANATICS_M\- I apply the Troop ID discount$/) do
+  idp_signin = window_opened_by { find("#imgMilitary").click }
+
+  within_window idp_signin do
+    sign_in_with_idme
+  end
+end
+
+Given(/^FANATICS_M\- I verify the Troop ID discount has been applied$/) do
+  expect(page).to have_text FigNewton.partners.fanatics.mobile.discount_copy
+  expect(page).to have_css(".couponAmount",:visible => true)
+
+  original_product_amt_string = find(".cartTotalAmount").text
+  actual_product_discounted_amt_string = find(".couponAmount").text
+  discount_applied = verify_discount(original_product_amt_string,actual_product_discounted_amt_string,".10")
+  expect(discount_applied).to be(true)
+end